refac: moved Converter
